+++ main.rb
@@ -1,7 +1,12 @@
+# Classes
 require_relative 'classes/person'
 require_relative 'classes/student'
 require_relative 'classes/teacher'
+
+# Decorators
+require_relative 'decorators/capitalize'
 require_relative 'decorators/nameable'
+require_relative 'decorators/trimmer'
 
 person = Person.new(22, 'maximilianus')
 person.correct_name
